+++ src/dispatch/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+
+angular.module('md.dispatch', ['ui.router']);
+
+require('./init');
+require('./services');
+require('./states');
+require('./controllers');